+++ .gitignore
@@ -5,3 +5,4 @@
 /myexpensesdb/nbproject/private/
 /myexpensesswing/build/
 /myexpensesdb/build/
+/myexpensesdb/dist/
