+++ MyExpenseClient/nbproject/private/private.properties
@@ -1,0 +1,4 @@
+deploy.ant.properties.file=/Users/sarkhanrasullu/Library/Application Support/NetBeans/8.2/tomcat90.properties
+j2ee.server.home=/Users/sarkhanrasullu/Workspace/apache-tomcat-9.0.5
+j2ee.server.instance=tomcat90:home=/Users/sarkhanrasullu/Workspace/apache-tomcat-9.0.5
+user.properties.file=/Users/sarkhanrasullu/Library/Application Support/NetBeans/8.2/build.properties
