+++ MyExpenseClient/web/login.html
@@ -1,0 +1,56 @@
+<!DOCTYPE html>
+<!-- html 5
+To change this license header, choose License Headers in Project Properties.
+To change this template file, choose Tools | Templates
+and open the template in the editor.
+-->
+<html>
+    <head>
+        <title>MyExpense.az</title>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="stylesheet" type="text/css" href="assets/myexpense.css">
+    </head>
+    <body>
+        <div>
+            <center>
+                <image id="logo" src="https://israff.com/image/logo.png"/>
+                <h3>Login</h3>
+                <table>
+                    <tbody>
+                        <tr>
+                            <td>
+                                <label for="username">username:</label>
+                            </td>
+                            <td>
+                                <input type="text" name="username" class="roundable"/>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td>
+                                <label for="parol">password:</label>
+                            </td>
+                            <td>
+                                <input type="password" name="parol" class="roundable"/>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td>
+                                <a href="https://www.google.com" target="_blank" class="myLink">Forgot password</a>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td>
+                                <input type="submit" value="Login" class="myBtn roundable"/>
+                            </td>
+                        </tr>
+
+                    </tbody>
+                </table>
+            </center>
+        </div>
+         
+    </body>
+</html>
+
+
