+++ MyExpenseClient/web/users.html
@@ -1,0 +1,113 @@
+<!DOCTYPE html>
+<!--
+To change this license header, choose License Headers in Project Properties.
+To change this template file, choose Tools | Templates
+and open the template in the editor.
+-->
+<html>
+    <head>
+        <title>Users</title>
+        <meta charset="UTF-8">
+    </head>
+    <body style="background-color: green">
+        <div>
+            <fieldset>
+                <legend>Fill fields:</legend>
+                <table>
+                    <tbody>
+                        <tr>
+                            <td>
+                                <label for="name" >
+                                    Name:
+                                </label>
+                            </td>
+                            <td>
+                                <input type="text" name="name"/>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td>
+                                <label for="surname">
+                                    Surname:
+                                </label>
+                            </td>
+                            <td>
+                                <input type="text" name="surname"/>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td>
+                                <label for="username">
+                                    Username:
+                                </label>
+                            </td>
+                            <td>
+                                <input type="text" name="username"/>
+                            </td>
+                        </tr>
+                        <tr>
+                            <td>
+                                <label for="roleid">
+                                    Role:
+                                </label>
+                            </td>
+                            <td>
+                                <select name="roleid">
+                                    <option>User</option>
+                                    <option>Admin</option>
+                                </select>
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
+                <br/>
+                <input type="submit" value="Search"/>
+                <input type="submit" value="Insert"/>
+                <input type="submit" value="Update"/>
+                <input type="submit" value="Delete"/>
+            </fieldset>
+        </div>
+        <hr/>
+        
+        <div>
+            <table border="1">
+                <thead>
+                    <tr>
+                        <th>id</th>
+                        <th>name</th>
+                        <th>surname</th>
+                        <th>username</th>
+                        <th>last login date</th>
+                        <th>registration date</th>
+                        <th>unregistration date</th>
+                        <th>role</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    <tr>
+                        <td>1</td>
+                        <td>Anar</td>
+                        <td>Salami</td>
+                        <td>AnarS</td>
+                        <td>10-10-18</td>
+                        <td>10-10-18</td>
+                        <td>10-10-18</td>
+                        <td>Admin</td>
+                    </tr>
+                    <tr>
+                        <td>2</td>
+                        <td>Anar</td>
+                        <td>Salami</td>
+                        <td>AnarS</td>
+                        <td>10-10-18</td>
+                        <td>10-10-18</td>
+                        <td>10-10-18</td>
+                        <td>Admin</td>
+                    </tr>
+                </tbody>
+            </table>
+
+        </div>
+        
+    </body>
+</html>
