+++ myexpensesdb/src/beans/UserRole.java
@@ -1,0 +1,25 @@
+package beans;
+
+public class UserRole {
+ 
+    private int id;//0
+    private String name; 
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+    
+    
+}
