+++ myexpensesdb/src/dao/impl/ExpenseDAOImpl.java
@@ -3,9 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dao;
+package dao.impl;
 
 import beans.Expense;
+import dao.inter.ExpenseDAOInter;
 import dbutility.DBUtility;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
