+++ myexpensesdb/src/dao/impl/UserDAOImpl.java
@@ -1,10 +1,12 @@
-package dao;
+package dao.impl;
 
 import beans.User;
+import dao.inter.UserDAOInter;
 import dbutility.DBUtility;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
@@ -24,7 +26,7 @@ public class UserDAOImpl extends DBUtility implements UserDAOInter {
             stmt.setString(2, user.getSurname());
             stmt.setString(3, user.getUsername());
             stmt.setString(4, user.getPassword());
-            stmt.setInt(5, user.getRoleId());
+            stmt.setInt(5, 2);
 
             int result = stmt.executeUpdate();
             if(result > 0) {
@@ -40,31 +42,6 @@ public class UserDAOImpl extends DBUtility implements UserDAOInter {
         }
         return 0;
     }
-    
-    
-    @Override
-   public int insertRole(User role) {
-        Connection connection = null;
-        try {
-            connection = connect();
-
-            String sql = "insert into user_role(name) values(?)";
-            PreparedStatement stmt = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-            stmt.setString(1, role.getRoleName());
-            int result = stmt.executeUpdate();
-            if(result > 0) {
-                int insertedId =  getInsertId(stmt);
-                role.setId(insertedId);
-                return insertedId;
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return -1;
-        }finally {
-            close(connection);
-        }
-        return 0;
-    } 
 
     @Override
     public User find(int id) {//jdbc specification-nin shertlerini odeyen kitabxana
@@ -149,26 +126,6 @@ public class UserDAOImpl extends DBUtility implements UserDAOInter {
             close(connection);
         }
     }
-    
-    @Override
-        public boolean deleteRole(int id) {
-        Connection connection = null;
-        try {
-            connection = connect();
-
-            String sql = "delete from user_role where id = ?";
-            PreparedStatement stmt = connection.prepareStatement(sql);
-            stmt.setInt(1, id);
-
-            stmt.executeUpdate();
-            return true;
-        } catch (Exception e) {
-            e.printStackTrace();
-            return false;
-        }finally {
-            close(connection);
-        }
-    }
 
     @Override
     public boolean update(int id, User user) {
@@ -181,27 +138,6 @@ public class UserDAOImpl extends DBUtility implements UserDAOInter {
             stmt.setString(1, user.getName());
             stmt.setString(2, user.getSurname());
             stmt.setInt(3, id);
-
-            stmt.executeUpdate();
-            return true;
-        } catch (Exception e) {
-            e.printStackTrace();
-            return false;
-        }finally {
-            close(connection);
-        }
-    }
-    
-    @Override
-        public boolean updateRole(int id, User role) {
-        Connection connection = null;
-        try {
-            connection = connect();
-
-            String sql = "update user_role set name = ? where id = ?";
-            PreparedStatement stmt = connection.prepareStatement(sql);
-            stmt.setString(1, role.getRoleName());
-            stmt.setInt(2, id);
 
             stmt.executeUpdate();
             return true;
@@ -248,36 +184,6 @@ public class UserDAOImpl extends DBUtility implements UserDAOInter {
         }
         return list;
     }
-    
-    @Override
-        public List<User> selectRoles() {
-        Connection connection = null;
-        List<User> list = new ArrayList<>();
-        try {
-            connection = connect();
-
-            String sql = "select * from user_role";
-            PreparedStatement stmt = connection.prepareStatement(sql);
-
-            ResultSet rs = stmt.executeQuery();
-            while(rs.next()){
-                int userId = rs.getInt("id");
-                String userName = rs.getString("name");
-
-                User user = new User();
-                user.setRoleId(userId);
-                user.setRoleName(userName);
-
-                list.add(user);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }finally {
-            close(connection);
-        }
-        return list;
-    }
-    
     public static int getInsertId(PreparedStatement stmt) throws Exception{
         ResultSet generatedKeys = stmt.getGeneratedKeys();
 
@@ -287,5 +193,68 @@ public class UserDAOImpl extends DBUtility implements UserDAOInter {
             return -1;
         }
     }
+@Override
+    public int logIn(String username, String password) {
+        Connection connection = null;
+        User user = new User();
+        
+        try {
+            connection = connect();
+            String sql = "SELECT * FROM user WHERE username = ? and password = ? ";
+            PreparedStatement stmt = connection.prepareStatement(sql);
+            stmt.setString(1, username);
+            stmt.setString(2, password);
 
+            ResultSet rs = stmt.executeQuery();
+
+            while (rs.next()) {
+                
+                user.setId(rs.getInt(1));
+                user.setName(rs.getString(2));
+                user.setSurname(rs.getString(3));
+                user.setUsername(rs.getString(4));
+                user.setPassword(rs.getString(5));
+                user.setRoleId(rs.getInt(9));
+               
+            }
+
+            return user.getRoleId();
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            close(connection);
+        }
+        return user.getRoleId();
+    }
+    @Override
+    public int signUp(User user) {//Ecemi
+        Connection connection = null;
+        try {
+            connection = connect();
+
+            String sql = "insert into user(name,surname,username,password,role_id, reg_date) values(?,?,?,?,?,now())";
+            PreparedStatement stmt = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
+            stmt.setString(1, user.getName());
+            stmt.setString(2, user.getSurname());
+            stmt.setString(3, user.getUsername());
+            stmt.setString(4, user.getPassword());
+            stmt.setInt(5, 2);
+
+            int result = stmt.executeUpdate();
+            if(result > 0) {
+                int insertedId =  getInsertId(stmt);
+                user.setId(insertedId);
+                return insertedId;
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return -1;
+        }finally {
+            close(connection);
+        }
+        return 0;
+    }
 }
