+++ myexpensesdb/src/dao/inter/ExpenseDAOInter.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dao;
+package dao.inter;
 
 import beans.Expense;
 import java.util.List;
