+++ myexpensesdb/src/dao/inter/ExpenseTypeDAOInter.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dao;
+package dao.inter;
 
 import beans.ExpenseType;
 import java.util.List;
