+++ myexpensesdb/src/dao/inter/UserRoleDAOInter.java
@@ -1,0 +1,17 @@
+package dao.inter;
+
+import beans.User;
+import beans.UserRole;
+import java.util.List;
+
+public interface UserRoleDAOInter {
+ 
+    public abstract List<UserRole> selectRoles();
+    
+    public abstract int insertRole(UserRole role);
+    
+    public abstract boolean deleteRole(int id);
+    
+    public abstract boolean updateRole (int id, UserRole role);
+    
+}
