+++ myexpensesdb/src/dbutility/DBConfig.java
@@ -6,7 +6,7 @@ class DBConfig {
     private static String port = "3306";
     private static String dbName = "myexpense";
     private static String username = "root";
-    private static String password = "6647";
+    private static String password = "12345";
 
     public static String getConnectionName(){
         String s = "jdbc:mysql://"+DBConfig.getIp()+":"+DBConfig.getPort()+"/"+DBConfig.getDbName()+"?useSSL=false";
